update size after leader transfer
